Refactor project structure and update CI/CD workflow
Removed old solution and project files (`csharp-ovh.sln`, `csharp-ovh.csproj`) and the previous GitHub Actions workflow file (`build-and-test.yml`). Introduced a new solution file (`ovh-api.sln`) and a new project file (`ovh-api.csproj`) as part of the project restructuring. Added a new GitHub Actions workflow (`build-test-publish.yml`) that not only builds and tests the project but also packages and publishes it to NuGet on push events to the master branch.