Refactored code for readability and updated API version
This commit includes a major code refactoring for improved readability and consistency. The `using` directives were rearranged and redundant ones were removed across multiple files. The visibility of several static variables was changed to `private static`. The `Assert.Equals` method was replaced with `Assert.That` in combination with `Is.EqualTo` for better readability and consistency across all test cases.

Several classes were updated to use the new API version (`ovh-eu-v1` instead of `ovh-eu`). The `Client` constructor and `TestCase` attributes were updated accordingly.

The `Client` class in `Client.cs` was significantly modified with the removal of several constants and the addition of new properties. The order of parameters in the `LoadConfiguration` method was changed and the `ExtractExceptionFromHttpResponse` method was updated to handle additional HTTP status codes.

The `ClientFactory.GetClient` method was removed in `GetRequests.cs` and `PostRequests.cs`. The `HttpResponseMessageFactory.Create` method was used to create `HttpResponseMessage` instances in `DeleteResponses.cs` and `GetResponses.cs`.

The `ovh-api.sln` file was modified to change the `Debug|Any CPU.ActiveCfg` and `Debug|Any CPU.Build.0` properties from `Release|Any CPU` to `Debug|Any CPU` for the `{FF3DE9E0-2E5B-4E39-B973-2FE50676D177}` project.

Finally, minor changes were made to improve the formatting of the code.